refactor(cli): extract replay catalog helpers
